+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Fixed
 
-- Header links are included in the extracted notes JetBrains/gradle-changelog-plugin#266
+- Header links are included in the extracted notes. JetBrains/gradle-changelog-plugin#266
+- Handle missing released items in `GetChangelogTask` gracefully. JetBrains/gradle-changelog-plugin#252
 
 ### Removed